requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
jinja2
python-multipart
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
